recorrect some codes

+++ src/components/ExpenseDate.js
@@ -1,8 +1,7 @@
-import React from 'react'
 
 const ExpenseDate = (props) => {
 
-    const month = props.date.toLocaleString('en-US', {month: 'long'});
+    const month = props.date.toLocaleString('en-US', {month: 'short'});
     const day   = props.date.toLocaleString('en-US', {day: '2-digit'});
     const year  = props.date.getFullYear();
